refactor: component index.ts
