+++ src/components/Card/Card.stories.tsx
@@ -1,0 +1,8 @@
+import { ComponentMeta } from '@storybook/react';
+
+import Card from './Card';
+
+export default {
+  title: 'Card',
+  component: Card,
+} as ComponentMeta<typeof Card>;
