+++ src/pages/CardAdded.tsx
@@ -1,8 +1,7 @@
 import { useContext, useRef } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import Button from '../components/Button';
-import Card from '../components/Card';
+import { Button, Card } from '../components';
 import { PaymentsContext } from '../context/PaymentsContext';
 
 const CardAdded = () => {
